Add AsyncController and use in place of OpenJ9-specific controller
