+++ dd-java-agent/agent-profiling/profiling-auxiliary/src/main/java/com/datadog/profiling/auxiliary/AuxiliaryImplementation.java
@@ -2,6 +2,7 @@ package com.datadog.profiling.auxiliary;
 
 import com.datadog.profiling.controller.OngoingRecording;
 import com.datadog.profiling.controller.RecordingData;
+import com.datadog.profiling.utils.ProfilingMode;
 import datadog.trace.bootstrap.config.provider.ConfigProvider;
 import java.util.Collections;
 import java.util.Set;
