+++ dd-java-agent/agent-profiling/profiling-auxiliary/src/main/java/com/datadog/profiling/auxiliary/AuxiliaryProfiler.java
@@ -1,6 +1,7 @@
 package com.datadog.profiling.auxiliary;
 
 import com.datadog.profiling.controller.OngoingRecording;
+import com.datadog.profiling.utils.ProfilingMode;
 import datadog.trace.api.config.ProfilingConfig;
 import datadog.trace.bootstrap.config.provider.ConfigProvider;
 import java.util.ServiceLoader;
