+++ dd-java-agent/agent-profiling/profiling-utils/src/test/resources/mockito-extensions/org.mockito.plugins.MockMaker
@@ -1,0 +1,1 @@
+mock-maker-inline
