+++ settings.gradle
@@ -47,16 +47,18 @@ include ':dd-java-agent:load-generator'
 
 // profiling
 include ':dd-java-agent:agent-profiling'
+include ':dd-java-agent:agent-profiling:profiling-async'
 include ':dd-java-agent:agent-profiling:profiling-auxiliary'
 include ':dd-java-agent:agent-profiling:profiling-auxiliary-async'
 include ':dd-java-agent:agent-profiling:profiling-context'
 include ':dd-java-agent:agent-profiling:profiling-controller'
 include ':dd-java-agent:agent-profiling:profiling-controller-jfr'
+include ':dd-java-agent:agent-profiling:profiling-controller-async'
 include ':dd-java-agent:agent-profiling:profiling-controller-openjdk'
 include ':dd-java-agent:agent-profiling:profiling-controller-oracle'
-include ':dd-java-agent:agent-profiling:profiling-controller-openj9'
 include ':dd-java-agent:agent-profiling:profiling-testing'
 include ':dd-java-agent:agent-profiling:profiling-uploader'
+include ':dd-java-agent:agent-profiling:profiling-utils'
 
 include ':communication'
